requirements.txt
tensorflow==2.4.1
pytorch==1.8.1
nltk==3.6.2
spacy==3.0.6
scikit-learn==0.24.2
flask==1.1.2
django==3.2.3
electron==11.4.3
sqlite3==3.35.4
psycopg2==2.8.6
mysql-connector-python==8.0.25
docker==5.0.0
boto3==1.17.84
azure==5.0.0
twilio==6.59.0
pyjwt==2.1.0
jenkins==1.7.0
gitlab==2.6.0
requests==2.25.1
numpy==1.20.3
pandas==1.2.4
beautifulsoup4==4.9.3
lxml==4.6.3
pytest==6.2.4
mock==4.0.3
coverage==5.5
flake8==3.9.2
black==21.5b2
mypy==0.812
isort==5.8.0
